Add Read instances and clean up the Show instance for SrcLoc.
The SrcLoc datatype is meant to tag source code locations in an abstract
syntax tree without affecting equality or ordering of AST nodes. The
Show instance for SrcLoc now always prints "noLoc", and the Read
instance accepts "noLoc" or "SrcLoc ..." where ... is the printed
representation of a Loc.

An alternative design is to have the Show instance for SrcLoc print the
empty string and have the Read instance consume no input and return
noLoc. The disadvantage of this approach is that the Show instances of
data types that contain values of type SrcLoc then output spurious spaces.